requirements.txt
datasets
matplotlib
sentence-transformers
scikit-learn
numpy
faiss-cpu
ipykernel